+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- **BREAKING** Prometheus ingester: Renamed `increase` query type to `counter_increase`.
+
+### Added
+- Prometheus ingester: New query type `histogram_increase` to generate events with for each bucket.
 
 ## [4.4.0] 2020-04-27
 ### Added